requirements.txt
python-nmap
pyfiglet
termcolor